fix delete_residues in chorizo (polymer) and mk_prepare_receptor
The list of residue IDs to delete is no longer passed to the init
of the chorizo object. Instead, it lives only on the constructors
from_prody and from_pdb_string. Then, the raw molecules passed to
the init of the chorizo no longer contains deleted residues.
In mk_prepare_receptor, deleted_residues are specified using the
same syntax as flexres.